refactor: one definition of the canonical Book Note shape
The migration was about to be undone by the next `libris add`.
create_book_note wrote "### Description" rather than the callout ADR 0009
settled on, minted no libris_id, and DEFAULT_FRONTMATTER carried neither
priority nor series. A note added the day after migrating would have differed
from every note the migration touched - the same drift this branch exists to
undo, starting over.

note_format.py now holds the canonical shape: which fields a Book Note carries
and in what order, how a body is written, and how a Libris ID is minted. Note
creation and the migration both render through it, so they cannot disagree.
DEFAULT_FRONTMATTER is derived from the field order rather than restated.

update_frontmatter_from_book was a second writer of the description section and
would have reintroduced the split on its own; it renders the callout too.

ensure_frontmatter_fields now mints a Libris ID when a note lacks one, so a
book typed straight into Obsidian gains an identity the next time clean or
autoenrich sees it, rather than staying unaddressable until the next migration.

Verified: a freshly created note is already migration-clean, and the migration
still reports no content lost on any of the 3,137 notes.

Refs #61.
